requirements.txt
fastapi
graphene
ariadne
motor
uvicorn
pika
python-dotenv
